Nonfunctional changes only

+++ container-search/src/main/java/com/yahoo/prelude/fastsearch/ByteField.java
@@ -24,7 +24,7 @@ public class ByteField extends DocsumField {
         if (value == EMPTY_VALUE) {
             return NanNumber.NaN;
         } else {
-            return Byte.valueOf(value);
+            return value;
         }
     }
